update webhook icon
